docs: Add comprehensive filter documentation and prompt templates
## Documentation Updates (CLAUDE.md)
- Added complete filter operator reference for all field types
- Documented 8 available prompt templates with descriptions
- Highlighted linked record filtering pitfall (has_any_of vs is)
- Updated PromptRegistry line count (213 → 447 lines)

## New Prompt Templates (4 added, total: 8)
1. filter_by_linked_record: Demonstrates has_any_of with record IDs
   - Includes warning about common "is" operator mistake
2. filter_by_numeric_range: Shows numeric comparisons
   - Supports optional min/max range filtering
3. filter_by_multiple_select: Tag/multi-select filtering
   - Supports has_any_of, has_all_of, is_exactly match types
4. filter_by_assigned_user: User field filtering
   - Includes reminder to use list_members for user IDs

## Filter Syntax Documentation
Added operator reference for all field types:
- Text-based: is, is_not, contains, not_contains, etc.
- Numeric: is_greater_than, is_equal_or_less_than, etc.
- Date: Special date object format with date_mode
- Linked Record: has_any_of, has_all_of, is_exactly (NOT "is")
- Multiple Select: has_any_of, has_all_of, is_exactly
- User fields: has_any_of with user ID arrays
- Files: file_name_contains, file_type_is
- Boolean: is, is_empty, is_not_empty

## Test Fix
- Fixed test_stats_tracker_tracks_calls to reset stats before testing
- Prevents interference from real-world testing sessions
- All 44 tests pass with 260 assertions

## Modular Architecture (Complete Refactor)
This commit includes the previously uncommitted refactoring work:

### MCP Protocol Layer (lib/smartsuite/mcp/)
- ToolRegistry: 15 tool schemas organized by category
- PromptRegistry: 8 prompt templates with full implementations
- ResourceRegistry: MCP resource listing (empty)

### API Client Layer (lib/smartsuite/api/)
- HttpClient: HTTP communication with auth and logging
- DataOperations: Solutions and tables management
- RecordOperations: Record CRUD with token optimization
- FieldOperations: Field CRUD operations
- MemberOperations: User and team management

### Formatters Layer (lib/smartsuite/formatters/)
- ResponseFormatter: Aggressive filtering for token optimization

### Code Reduction
- smartsuite_server.rb: 769 → 262 lines (66% reduction)
- smartsuite_client.rb: 708 → 30 lines (96% reduction)
- Total: 9 new focused modules vs 2 monolithic files
