+++ smartsuite_server.rb
@@ -3,6 +3,9 @@
 require 'json'
 require_relative 'lib/smartsuite_client'
 require_relative 'lib/api_stats_tracker'
+require_relative 'lib/smartsuite/mcp/tool_registry'
+require_relative 'lib/smartsuite/mcp/prompt_registry'
+require_relative 'lib/smartsuite/mcp/resource_registry'
 
 class SmartSuiteServer
   def initialize
@@ -132,529 +135,19 @@ class SmartSuiteServer
   end
 
   def handle_prompts_list(request)
-    {
-      'jsonrpc' => '2.0',
-      'id' => request['id'],
-      'result' => {
-        'prompts' => [
-          {
-            'name' => 'filter_active_records',
-            'description' => 'Example: Filter records where status is "active"',
-            'arguments' => [
-              {
-                'name' => 'table_id',
-                'description' => 'The table ID to query',
-                'required' => true
-              },
-              {
-                'name' => 'status_field',
-                'description' => 'The field slug for status (default: "status")',
-                'required' => false
-              },
-              {
-                'name' => 'fields',
-                'description' => 'Comma-separated list of field slugs to return',
-                'required' => true
-              }
-            ]
-          },
-          {
-            'name' => 'filter_by_date_range',
-            'description' => 'Example: Filter records within a date range',
-            'arguments' => [
-              {
-                'name' => 'table_id',
-                'description' => 'The table ID to query',
-                'required' => true
-              },
-              {
-                'name' => 'date_field',
-                'description' => 'The field slug for the date field',
-                'required' => true
-              },
-              {
-                'name' => 'start_date',
-                'description' => 'Start date (YYYY-MM-DD format)',
-                'required' => true
-              },
-              {
-                'name' => 'end_date',
-                'description' => 'End date (YYYY-MM-DD format)',
-                'required' => true
-              },
-              {
-                'name' => 'fields',
-                'description' => 'Comma-separated list of field slugs to return',
-                'required' => true
-              }
-            ]
-          },
-          {
-            'name' => 'list_tables_by_solution',
-            'description' => 'Example: List all tables in a specific solution',
-            'arguments' => [
-              {
-                'name' => 'solution_id',
-                'description' => 'The solution ID to filter by',
-                'required' => true
-              }
-            ]
-          },
-          {
-            'name' => 'filter_records_contains_text',
-            'description' => 'Example: Filter records where a field contains specific text',
-            'arguments' => [
-              {
-                'name' => 'table_id',
-                'description' => 'The table ID to query',
-                'required' => true
-              },
-              {
-                'name' => 'field_slug',
-                'description' => 'The field slug to search in',
-                'required' => true
-              },
-              {
-                'name' => 'search_text',
-                'description' => 'The text to search for',
-                'required' => true
-              },
-              {
-                'name' => 'fields',
-                'description' => 'Comma-separated list of field slugs to return',
-                'required' => true
-              }
-            ]
-          }
-        ]
-      }
-    }
+    SmartSuite::MCP::PromptRegistry.prompts_list(request)
   end
 
   def handle_prompt_get(request)
-    prompt_name = request.dig('params', 'name')
-    arguments = request.dig('params', 'arguments') || {}
-
-    prompt_text = case prompt_name
-    when 'filter_active_records'
-      status_field = arguments['status_field'] || 'status'
-      fields = arguments['fields']&.split(',')&.map(&:strip) || []
-
-      "Use the list_records tool with these parameters:\n\n" +
-      "table_id: #{arguments['table_id']}\n" +
-      "fields: #{fields.inspect}\n" +
-      "filter: {\n" +
-      "  \"operator\": \"and\",\n" +
-      "  \"fields\": [\n" +
-      "    {\n" +
-      "      \"field\": \"#{status_field}\",\n" +
-      "      \"comparison\": \"is\",\n" +
-      "      \"value\": \"active\"\n" +
-      "    }\n" +
-      "  ]\n" +
-      "}"
-
-    when 'filter_by_date_range'
-      date_field = arguments['date_field']
-      start_date = arguments['start_date']
-      end_date = arguments['end_date']
-      fields = arguments['fields']&.split(',')&.map(&:strip) || []
-
-      "Use the list_records tool with these parameters:\n\n" +
-      "table_id: #{arguments['table_id']}\n" +
-      "fields: #{fields.inspect}\n" +
-      "filter: {\n" +
-      "  \"operator\": \"and\",\n" +
-      "  \"fields\": [\n" +
-      "    {\n" +
-      "      \"field\": \"#{date_field}\",\n" +
-      "      \"comparison\": \"is_after\",\n" +
-      "      \"value\": {\n" +
-      "        \"date_mode\": \"exact_date\",\n" +
-      "        \"date_mode_value\": \"#{start_date}\"\n" +
-      "      }\n" +
-      "    },\n" +
-      "    {\n" +
-      "      \"field\": \"#{date_field}\",\n" +
-      "      \"comparison\": \"is_before\",\n" +
-      "      \"value\": {\n" +
-      "        \"date_mode\": \"exact_date\",\n" +
-      "        \"date_mode_value\": \"#{end_date}\"\n" +
-      "      }\n" +
-      "    }\n" +
-      "  ]\n" +
-      "}"
-
-    when 'list_tables_by_solution'
-      "Use the list_tables tool with this parameter:\n\n" +
-      "solution_id: #{arguments['solution_id']}\n\n" +
-      "This will return only tables from the specified solution."
-
-    when 'filter_records_contains_text'
-      field_slug = arguments['field_slug']
-      search_text = arguments['search_text']
-      fields = arguments['fields']&.split(',')&.map(&:strip) || []
-
-      "Use the list_records tool with these parameters:\n\n" +
-      "table_id: #{arguments['table_id']}\n" +
-      "fields: #{fields.inspect}\n" +
-      "filter: {\n" +
-      "  \"operator\": \"and\",\n" +
-      "  \"fields\": [\n" +
-      "    {\n" +
-      "      \"field\": \"#{field_slug}\",\n" +
-      "      \"comparison\": \"contains\",\n" +
-      "      \"value\": \"#{search_text}\"\n" +
-      "    }\n" +
-      "  ]\n" +
-      "}"
-
-    else
-      "Unknown prompt: #{prompt_name}"
-    end
-
-    {
-      'jsonrpc' => '2.0',
-      'id' => request['id'],
-      'result' => {
-        'messages' => [
-          {
-            'role' => 'user',
-            'content' => {
-              'type' => 'text',
-              'text' => prompt_text
-            }
-          }
-        ]
-      }
-    }
+    SmartSuite::MCP::PromptRegistry.prompt_get(request)
   end
 
   def handle_resources_list(request)
-    {
-      'jsonrpc' => '2.0',
-      'id' => request['id'],
-      'result' => {
-        'resources' => []
-      }
-    }
+    SmartSuite::MCP::ResourceRegistry.resources_list(request)
   end
 
   def handle_tools_list(request)
-    {
-      'jsonrpc' => '2.0',
-      'id' => request['id'],
-      'result' => {
-        'tools' => [
-          {
-            'name' => 'list_solutions',
-            'description' => 'List all solutions in your SmartSuite workspace (solutions contain tables)',
-            'inputSchema' => {
-              'type' => 'object',
-              'properties' => {},
-              'required' => []
-            }
-          },
-          {
-            'name' => 'list_members',
-            'description' => 'List all members (users) in your SmartSuite workspace. Use this to get user IDs for assigning people to records. Optionally filter by solution_id to only show members who have access to that solution (saves tokens).',
-            'inputSchema' => {
-              'type' => 'object',
-              'properties' => {
-                'limit' => {
-                  'type' => 'number',
-                  'description' => 'Maximum number of members to return (default: 100). Ignored when solution_id is provided.'
-                },
-                'offset' => {
-                  'type' => 'number',
-                  'description' => 'Number of members to skip (for pagination). Ignored when solution_id is provided.'
-                },
-                'solution_id' => {
-                  'type' => 'string',
-                  'description' => 'Optional: Filter members by solution ID. Returns only members who have access to this solution. This saves tokens by filtering server-side.'
-                }
-              },
-              'required' => []
-            }
-          },
-          {
-            'name' => 'list_tables',
-            'description' => 'List all tables (apps) in your SmartSuite workspace. Optionally filter by solution_id to only show tables from a specific solution.',
-            'inputSchema' => {
-              'type' => 'object',
-              'properties' => {
-                'solution_id' => {
-                  'type' => 'string',
-                  'description' => 'Optional: Filter tables by solution ID. Use list_solutions first to get solution IDs.'
-                }
-              },
-              'required' => []
-            }
-          },
-          {
-            'name' => 'get_table',
-            'description' => 'Get a specific table by ID including its structure (fields, their slugs, types, etc). Use this BEFORE querying records to understand what fields are available for filtering and selection.',
-            'inputSchema' => {
-              'type' => 'object',
-              'properties' => {
-                'table_id' => {
-                  'type' => 'string',
-                  'description' => 'The ID of the table to retrieve'
-                }
-              },
-              'required' => ['table_id']
-            }
-          },
-          {
-            'name' => 'list_records',
-            'description' => 'List records from a SmartSuite table. DEFAULT: Returns only id + title for minimal context usage. Use fields parameter for specific data or summary_only for statistics.',
-            'inputSchema' => {
-              'type' => 'object',
-              'properties' => {
-                'table_id' => {
-                  'type' => 'string',
-                  'description' => 'The ID of the table to query'
-                },
-                'limit' => {
-                  'type' => 'number',
-                  'description' => 'Maximum number of records to return (default: 5 for minimal context usage)'
-                },
-                'offset' => {
-                  'type' => 'number',
-                  'description' => 'Number of records to skip (for pagination)'
-                },
-                'filter' => {
-                  'type' => 'object',
-                  'description' => 'Filter criteria. STRUCTURE: {"operator": "and|or", "fields": [{"field": "field_slug", "comparison": "operator", "value": "value"}]}. EXAMPLES: 1) Single filter: {"operator": "and", "fields": [{"field": "status", "comparison": "is", "value": "active"}]}. 2) Multiple filters: {"operator": "and", "fields": [{"field": "status", "comparison": "is", "value": "active"}, {"field": "priority", "comparison": "is_greater_than", "value": 3}]}. 3) Date filter (IMPORTANT - use date value object): {"operator": "and", "fields": [{"field": "due_date", "comparison": "is_after", "value": {"date_mode": "exact_date", "date_mode_value": "2025-01-01"}}]}. OPERATORS: is, is_not, contains, is_greater_than, is_less_than, is_empty, is_not_empty, is_before, is_after. NOTE: Date fields require value as object with date_mode and date_mode_value.'
-                },
-                'sort' => {
-                  'type' => 'array',
-                  'description' => 'Sort criteria as array of field-direction pairs. Example: [{"field": "created_on", "direction": "desc"}]',
-                  'items' => {
-                    'type' => 'object',
-                    'properties' => {
-                      'field' => {
-                        'type' => 'string',
-                        'description' => 'Field slug to sort by'
-                      },
-                      'direction' => {
-                        'type' => 'string',
-                        'description' => 'Sort direction: "asc" or "desc"',
-                        'enum' => ['asc', 'desc']
-                      }
-                    }
-                  }
-                },
-                'fields' => {
-                  'type' => 'array',
-                  'description' => 'Optional: Specific field slugs to return. Default returns only id + title. Specify fields to get additional data.',
-                  'items' => {
-                    'type' => 'string'
-                  }
-                },
-                'summary_only' => {
-                  'type' => 'boolean',
-                  'description' => 'If true, returns statistics/summary instead of actual records. Minimal context usage for overview purposes.'
-                },
-                'full_content' => {
-                  'type' => 'boolean',
-                  'description' => 'If true, returns full field content without truncation. Default (false): strings truncated to 500 chars. Use when you need complete field values (like full descriptions) to avoid multiple get_record calls.'
-                }
-              },
-              'required' => ['table_id']
-            }
-          },
-          {
-            'name' => 'get_record',
-            'description' => 'Get a specific record by ID from a SmartSuite table',
-            'inputSchema' => {
-              'type' => 'object',
-              'properties' => {
-                'table_id' => {
-                  'type' => 'string',
-                  'description' => 'The ID of the table'
-                },
-                'record_id' => {
-                  'type' => 'string',
-                  'description' => 'The ID of the record to retrieve'
-                }
-              },
-              'required' => ['table_id', 'record_id']
-            }
-          },
-          {
-            'name' => 'create_record',
-            'description' => 'Create a new record in a SmartSuite table',
-            'inputSchema' => {
-              'type' => 'object',
-              'properties' => {
-                'table_id' => {
-                  'type' => 'string',
-                  'description' => 'The ID of the table'
-                },
-                'data' => {
-                  'type' => 'object',
-                  'description' => 'The record data as key-value pairs (field_slug: value)'
-                }
-              },
-              'required' => ['table_id', 'data']
-            }
-          },
-          {
-            'name' => 'update_record',
-            'description' => 'Update an existing record in a SmartSuite table',
-            'inputSchema' => {
-              'type' => 'object',
-              'properties' => {
-                'table_id' => {
-                  'type' => 'string',
-                  'description' => 'The ID of the table'
-                },
-                'record_id' => {
-                  'type' => 'string',
-                  'description' => 'The ID of the record to update'
-                },
-                'data' => {
-                  'type' => 'object',
-                  'description' => 'The record data to update as key-value pairs (field_slug: value)'
-                }
-              },
-              'required' => ['table_id', 'record_id', 'data']
-            }
-          },
-          {
-            'name' => 'delete_record',
-            'description' => 'Delete a record from a SmartSuite table',
-            'inputSchema' => {
-              'type' => 'object',
-              'properties' => {
-                'table_id' => {
-                  'type' => 'string',
-                  'description' => 'The ID of the table'
-                },
-                'record_id' => {
-                  'type' => 'string',
-                  'description' => 'The ID of the record to delete'
-                }
-              },
-              'required' => ['table_id', 'record_id']
-            }
-          },
-          {
-            'name' => 'add_field',
-            'description' => 'Add a new field to a SmartSuite table. Returns the created field object.',
-            'inputSchema' => {
-              'type' => 'object',
-              'properties' => {
-                'table_id' => {
-                  'type' => 'string',
-                  'description' => 'The ID of the table to add the field to'
-                },
-                'field_data' => {
-                  'type' => 'object',
-                  'description' => 'Field configuration object with slug, label, field_type, and params. Example: {"slug": "abc123defg", "label": "My Field", "field_type": "textfield", "params": {"help_text": "Enter text"}, "is_new": true}'
-                },
-                'field_position' => {
-                  'type' => 'object',
-                  'description' => 'Optional: Position metadata. Example: {"prev_sibling_slug": "field_slug"} to place after another field'
-                },
-                'auto_fill_structure_layout' => {
-                  'type' => 'boolean',
-                  'description' => 'Optional: Enable automatic layout structure updates (default: true)'
-                }
-              },
-              'required' => ['table_id', 'field_data']
-            }
-          },
-          {
-            'name' => 'bulk_add_fields',
-            'description' => 'Add multiple fields to a SmartSuite table in one request. Note: Certain field types are not supported in bulk operations (e.g., Formula, Count, TimeTracking).',
-            'inputSchema' => {
-              'type' => 'object',
-              'properties' => {
-                'table_id' => {
-                  'type' => 'string',
-                  'description' => 'The ID of the table to add fields to'
-                },
-                'fields' => {
-                  'type' => 'array',
-                  'description' => 'Array of field configuration objects. Each should have slug, label, field_type, icon, params, and is_new.',
-                  'items' => {
-                    'type' => 'object'
-                  }
-                },
-                'set_as_visible_fields_in_reports' => {
-                  'type' => 'array',
-                  'description' => 'Optional: Array of view (report) IDs where the added fields should be visible',
-                  'items' => {
-                    'type' => 'string'
-                  }
-                }
-              },
-              'required' => ['table_id', 'fields']
-            }
-          },
-          {
-            'name' => 'update_field',
-            'description' => 'Update an existing field in a SmartSuite table. Returns the updated field object.',
-            'inputSchema' => {
-              'type' => 'object',
-              'properties' => {
-                'table_id' => {
-                  'type' => 'string',
-                  'description' => 'The ID of the table containing the field'
-                },
-                'slug' => {
-                  'type' => 'string',
-                  'description' => 'The slug of the field to update'
-                },
-                'field_data' => {
-                  'type' => 'object',
-                  'description' => 'Updated field configuration object with label, field_type, and params. Example: {"label": "Updated Label", "field_type": "textfield", "params": {"help_text": "New help text"}}'
-                }
-              },
-              'required' => ['table_id', 'slug', 'field_data']
-            }
-          },
-          {
-            'name' => 'delete_field',
-            'description' => 'Delete a field from a SmartSuite table. Returns the deleted field object.',
-            'inputSchema' => {
-              'type' => 'object',
-              'properties' => {
-                'table_id' => {
-                  'type' => 'string',
-                  'description' => 'The ID of the table containing the field'
-                },
-                'slug' => {
-                  'type' => 'string',
-                  'description' => 'The slug of the field to delete'
-                }
-              },
-              'required' => ['table_id', 'slug']
-            }
-          },
-          {
-            'name' => 'get_api_stats',
-            'description' => 'Get API call statistics tracked by user, solution, table, and HTTP method',
-            'inputSchema' => {
-              'type' => 'object',
-              'properties' => {},
-              'required' => []
-            }
-          },
-          {
-            'name' => 'reset_api_stats',
-            'description' => 'Reset all API call statistics',
-            'inputSchema' => {
-              'type' => 'object',
-              'properties' => {},
-              'required' => []
-            }
-          }
-        ]
-      }
-    }
+    SmartSuite::MCP::ToolRegistry.tools_list(request)
   end
 
   def handle_tool_call(request)
